+++ src/spaces/discrete.jl
@@ -11,7 +11,7 @@ agents are stored in a field `stored_ids` of the space.
 =#
 
 export positions, npositions, ids_in_position, agents_in_position,
-       empty_positions, random_empty, has_empty_positions
+       empty_positions, random_empty, has_empty_positions, empty_nearby_positions
 
 
 positions(model::ABM) = positions(model.space)
@@ -116,6 +116,27 @@ function random_empty(model::ABM{<:DiscreteSpace}, cutoff = 0.998)
         return rand(model.rng, collect(empty))
     end
 end
+
+"""
+    empty_nearby_positions(position, model::ABM{<:DiscreteSpace}, r = 1; kwargs...)
+    empty_nearby_positions(agent, model::ABM{<:DiscreteSpace}, r = 1; kwargs...)
+
+Return an iterable of all empty positions within "radius" `r` of the given `position`
+or the position of an `agent::AbstractAgent`.
+
+The value of `r` and possible keywords operate identically to [`nearby_positions`](@ref).
+
+This function only exists for discrete spaces with a finite amount of positions.
+"""
+function empty_nearby_positions(agent::AbstractAgent, model, r = 1; kwargs...)
+    return empty_nearby_positions(agent.pos, model, r; kwargs...)
+end
+function empty_nearby_positions(pos, model, r = 1; kwargs...)
+    return Iterators.filter(
+        Base.Fix2(isempty, model), nearby_positions(pos, model, r; kwargs...)
+    )
+end
+
 
 #######################################################################################
 # Discrete space extra agent adding stuff